# openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "CRUX API",
    "version": "1.0.0",
    "description": "ComfHutt CRUX Property Intelligence Engine API"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    },
    {
      "url": "https://api.crux.comfhutt.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "ClerkAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Clerk session token"
      }
    },
    "schemas": {
...
      "ScoreRequest": {
        "type": "object",
        "required": [
          "propertyAddress",
          "city",
          "pincode",
          "lifecycleStage",
          "investorIntent"
        ],
        "properties": {
          "propertyAddress": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "pincode": {
            "type": "string"
          },
          "lifecycleStage": {
            "type": "string",
            "enum": [
              "near_completion",
              "delivered",
              "established",
              "mature"
            ]
          },
          "investorIntent": {
            "type": "string",
            "enum": [
              "yield",
              "appreciation",
              "balanced"
            ]
          },
          "lat": {
            "type": "number"
          },
          "lng": {
            "type": "number"
          }
        }
      },
      "ParameterScore": {
        "type": "object",
        "required": [
          "name",
          "score",
          "weight",
          "source",
          "confidence",
          "lastUpdated"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "weight": {
            "type": "number"
          },
          "source": {
            "type": "string"
          },
          "confidence": {
            "type": "number"
          },
          "lastUpdated": {
            "type": "string"
          }
        }
      },
      "ScoreCategory": {
        "type": "object",
        "required": [
          "name",
          "score",
          "parameters"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "score": {
            "type": "number"
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParameterScore"
            }
          }
        }
      },
...
      "MarketValuation": {
        "type": "object",
        "required": [
          "weightedFairValueMin",
          "weightedFairValueMax",
          "methodVarianceFlag",
          "methods"
        ],
        "properties": {
          "weightedFairValueMin": {
            "type": "number"
          },
          "weightedFairValueMax": {
            "type": "number"
          },
          "methodVarianceFlag": {
            "type": "boolean"
          },
          "methods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValuationMethod"
            }
          },
          "vsListedPricePercent": {
            "type": "number"
          }
        }
      },
      "CruxScoreResponse": {
        "type": "object",
        "required": [
          "scoreId",
          "shareToken",
          "compositeScore",
          "macroMarketCycle",
          "categories",
          "valuation",
          "reportSummary",
          "confidenceScore",
          "methodologyVersion",
          "createdAt"
        ],
        "properties": {
          "scoreId": {
            "type": "string"
          },
          "shareToken": {
            "type": "string"
          },
          "compositeScore": {
            "type": "number"
          },
          "macroMarketCycle": {
            "type": "string",
            "enum": [
              "growth",
              "consolidation",
              "correction",
              "recovery"
            ]
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScoreCategory"
            }
          },
          "valuation": {
            "$ref": "#/components/schemas/MarketValuation"
          },
          "reportSummary": {
            "type": "string"
          },
          "confidenceScore": {
            "type": "number"
          },
          "methodologyVersion": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        }
      },
...
      "CruxReportResponse": {
        "type": "object",
        "required": [
          "id",
          "property_id",
          "score_id",
          "intent_profile",
          "summary",
          "category_narratives",
          "risk_flags",
          "positive_signals",
          "research_highlights",
          "citations",
          "sebi_disclaimer",
          "crux_version",
          "generated_at",
          "ttl_expires_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "property_id": {
            "type": "string"
          },
          "score_id": {
            "type": "string"
          },
          "intent_profile": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "category_narratives": {
            "type": "object",
            "required": [
              "legal_title",
              "location_quality",
              "developer_reliability",
              "market_valuation",
              "demand_signals"
            ],
            "properties": {
              "legal_title": {
                "type": "string"
              },
              "location_quality": {
                "type": "string"
              },
              "developer_reliability": {
                "type": "string"
              },
              "market_valuation": {
                "type": "string"
              },
              "demand_signals": {
                "type": "string"
              }
            }
          },
          "risk_flags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "positive_signals": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "research_highlights": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "citations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResearchCitation"
            }
          },
          "sebi_disclaimer": {
            "type": "string"
          },
          "crux_version": {
            "type": "string"
          },
          "generated_at": {
            "type": "string"
          },
          "ttl_expires_at": {
            "type": "string"
          }
        }
      },
      "ResearchCitation": {
        "type": "object",
        "required": [
          "claim",
          "source_title",
          "source_url_or_path",
          "authority_tier",
          "observed_at"
        ],
        "properties": {
          "claim": {
            "type": "string"
          },
          "source_title": {
            "type": "string"
          },
          "source_url_or_path": {
            "type": "string"
          },
          "authority_tier": {
            "type": "string",
            "enum": [
              "official",
              "primary",
              "secondary",
              "unknown"
            ]
          },
          "observed_at": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "ResearchRun": {
        "type": "object",
        "required": [
          "id",
          "property_id",
          "status",
          "initiated_by_surface",
          "provider",
          "seed_urls",
          "document_paths",
          "summary_counts",
          "started_at",
          "ttl_expires_at",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "property_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "success",
              "partial_failed",
              "failed"
            ]
          },
          "initiated_by_surface": {
            "type": "string",
            "enum": [
              "api",
              "lens",
              "report"
            ]
          },
          "provider": {
            "type": "string",
            "enum": [
              "tavily"
            ]
          },
          "seed_urls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "document_paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "summary_counts": {
            "type": "object"
          },
          "started_at": {
            "type": "string"
          },
          "completed_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "ttl_expires_at": {
            "type": "string"
          },
          "last_error": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string"
          }
        }
      },
      "ResearchRequest": {
        "type": "object",
        "properties": {
          "seed_urls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "document_paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "force_refresh": {
            "type": "boolean"
          },
          "surface": {
            "type": "string",
            "enum": [
              "api",
              "lens",
              "report"
            ]
          }
        }
      },
      "ResearchRunResponse": {
        "type": "object",
        "required": [
          "run",
          "digest",
          "reused_cache"
        ],
        "properties": {
          "run": {
            "$ref": "#/components/schemas/ResearchRun"
          },
          "digest": {
            "$ref": "#/components/schemas/ResearchEvidenceDigest"
          },
          "reused_cache": {
            "type": "boolean"
          }
        }
      },
      "CardShareResponse": {
        "type": "object",
        "required": [
          "shareToken",
          "shareUrl",
          "ogTitle",
          "ogDescription",
          "compositeScore",
          "propertyAddress",
          "expiresAt"
        ],
        "properties": {
          "shareToken": {
            "type": "string"
          },
          "shareUrl": {
            "type": "string"
          },
          "ogTitle": {
            "type": "string"
          },
          "ogDescription": {
            "type": "string"
          },
          "compositeScore": {
            "type": "number"
          },
          "propertyAddress": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        }
      }
    }
  },
  "security": [
    {
      "ClerkAuth": []
    }
  ],
  "paths": {
...
    "/crux/score": {
      "post": {
        "operationId": "createScore",
        "tags": [
          "CRUX"
        ],
        "summary": "Score a property across 20+ parameters",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CruxScoreResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Rate limited"
          }
        }
      }
    },
...
    "/crux/report/{property_id}": {
      "get": {
        "operationId": "getReport",
        "tags": [
          "CRUX"
        ],
        "summary": "Get full CRUX report for a property",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CruxReportResponse"
                }
              }
            }
          }
        }
      }
    },
    "/crux/research/{property_id}": {
      "get": {
        "operationId": "getResearch",
        "tags": [
          "CRUX"
        ],
        "summary": "Get the latest research run and accepted evidence for a property",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No research run found"
          }
        }
      },
      "post": {
        "operationId": "runResearch",
        "tags": [
          "CRUX"
        ],
        "summary": "Run or refresh research evidence for a property",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/crux/card/{shareToken}": {
      "get": {
        "operationId": "getShareCard",
        "tags": [
          "CRUX"
        ],
        "security": [],
        "summary": "Get public share card data (no auth required)",
        "parameters": [
          {
            "in": "path",
            "name": "shareToken",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardShareResponse"
                }
              }
            }
          },
          "404": {
            "description": "Card not found or expired"
          }
        }
      }
    }
  }
}